requirements.txt
streamlit
python-dotenv
azure-ai-formrecognizer